requirements.txt
requests
pymongo
python-dotenv
networkx
matplotlib
pandas
